Updates from Rijk
Updated:

I don't think this div is ending anywhere. Also I don't think you need a top-level container div here: you can just apply those styles to the body instead.

There is a glyph for №, but https://github.com/lucytrep/manuscript/blob/941749ba21690d050b8086e4e7d2d5024a38ebd1/index.html#L11is a good workaround in case the typeface used doesn't have it

An article publish date is a great opportunity for a time

The code here is now part of the actual content of the heading, and will be read as such in screenreaders and other tools. In this particular case, I believe the code snippets are just intended as visual flair, not as content, in which case CSS is the better place for them. This is a perfect use case for ::before and ::after on h3

p.code makes me think this should probably just be a code

Content of a blockquote should be a paragraph